should be called for-of
