+++ test/stache-test.js
@@ -2,7 +2,7 @@
 require('./expression-test');
 require('../helpers/-debugger-test');
 require('./nodelist-test');
-require('../helpers/-for-in-test');
+require('../helpers/-for-of-test');
 require('../helpers/-each-test');
 require('./section-test');
 var stache = require('../can-stache');
